Issue #4135: Radix Window Masking
Replace the column-by-column partition and peer mask construction
with a single scan over the radix data using the radix scanning
iterators from IEJoin.
